Changed site structure.
Based on feedback from Shaun, I've tried to make the site structure
reflect the use-cases of people visiting it. Instead of having the
list of projects at the top level, I've grouped them under headings
for learning about the project in general, getting involved with a
project with a specific set of features and finding documentaion
about a specific project.