Added links to projects home pages.
Linked directly to a (new) project property, 'home', for each
projects. The link appears in the header drop-down for the project
and also in the footer of each page in the project docs.

Also reorganised the nav bar so that links to all the pages within
a project are accessible there, instead of having to rely on any
intra-documentation links in the text itself.

This is trying to fix issue #9.